Put change password and sign out into navbar
Refactored the change password and sign out functionality into the
navbar. Made it so that sign up/in shows when the user is not logged
in but is hidden when logged in, and change password and sign out
are hidden if not logged in but show when logged in. Tested the
functionality of the navbar and everything works correctly.

Reorganized index.html to start with a good base to index products.
NextSteps: Make HandleBars template for showing products.