Create cart events, ui, and api chain
Team created a cart events, ui, and api to get form data when
"Add to Cart" is pressed on a product. Successfully tested that
getFormFields gets the data out of dropdown inputs correctly. Will need
to create "hidden" form values with product info to get the rest of
the product info into the object to post to the user's cart.
We are totally bosses.
NextSteps: Create hiddine form values and update user cart.