Create cui updateCartSuccess
Created the success state for the onUpdateCart event. Tested just by
console logging the response from the backend, which returns the entire
user object successfully. Also checked the backend by opening a mongo
shell and checking our test user; it adds the item correctly.
Next Steps: Update the cart of the user data on the frontend