Reconfigure webpack for less/bootstrap-less
The template we want to use is only accessible using less or css.
I reconfigured webpack to use less via node modules (less, less-loader
and bootstrap-less). The bootstrap styles are loading as before and
the some of the template styling is coming through.
Next step: I will go through the template to see what classes and
are being used to style the css and attempt to apply them to our html
as it is.