Rework show cart handlebars template to restyle
I redid the hb template to conform to the new style
It renders the cart in a table and the cart items
are listed as table rows.
Next step: need to add update cart button to remove
itens in cart. Will probably need to store values
in the the tables all within a form.