Clean up navbar and re-format modals
We removed the extra navbar at the top of the page once we got the
current navbar working. We re-formatted the form body in the modals
to center it and allow for more room for inputs. We also changed some
of the styling of the body to remove extraneous white space.