Get remove CartProduct working
Got the DELETE action for CartProduct working. Had to modify the event
chain for onDestroyCartProduct slightly, and add a data-id attribute
to the table data row for the remove button of the cart handlebar
template that stores the id of the CartProduct. Also had a rebase
conflict when I tried to apply my stashed changes on top of a rebase
which we fixed.
Next Steps: Get PATCH for CartProduct working, quantity and size.