Fix sign in bug that prevents user adding to cart
Fixed a bug that prevented a user from adding a product to their cart
if they signed in while they were viewing a single product page. Now
if a user signs in while on a product page the Add to Cart button
updates and allows the user to add a product to their cart.