Mobile styling of product imaging
Adjusted the padding and element positions on the items page below
sm col width. Pulls add to cart button left and adds 15px padding around
all page elements previously touching the edges of the screen. Moved the
home button so it was positioned nicer in relation to the add to
cart button.