Fix dropdown value clearing on Sign In
Fixed an issue where the dropdowns for Size and Quantity on the
show-product.handlebars page were getting their values cleared
if a user signed in while viewing the page.
I just commented out the line in auth/ui.js because I'm not sure
if it is needed for something else but everything seems to be
working correctly without it.